requirements/training.txt
numpy>=1.21.0
scipy>=1.7.0
torch>=1.11.0
torchaudio>=0.11.0
pytorch-lightning>=1.6.0
matplotlib>=3.5.0
librosa>=0.9.0
soundfile>=0.10.0
pyroomacoustics>=0.6.0
tqdm>=4.64.0
wandb>=0.15.0    # Optional for logging and experiment tracking
pytorch-lightning
google-cloud